Add Ketrew-backend library
This comes with a Ketrew binary linked with the plugin (one does not
need dynamic linking), and a simple workflow to test it.

The HTTP work is in the new `Client` module which is usable by
itself (cf. client-server test).

The configuration in `src/test/ketrew_env_config.ml` is adapted from the
one for the Docker container of Ketrew.

Here is how a I run my tests, and `mk` script that runs shell functions,
and puts both servers in a `screen` sesssion:

```shell
kserver () {
    KETREW_CONFIG=src/test/ketrew_env_config.ml KETREW_PROFILE=server \
                 PORT=8080 AUTH_TOKEN=coclobas \
                 ./coclobas-ketrew.byte $*
}
kgui (){
    open http://localhost:8080/gui?token=coclobas
}
ksubmit () {
    KETREW_CONFIG=src/test/ketrew_env_config.ml KETREW_PROFILE=client \
                 PORT=8080 AUTH_TOKEN=coclobas ./coclobas-ketrew-workflow-test.byte $*
}

screentest () {
    cat ~/.screenrc > /tmp/screencoclorc
    cat >> /tmp/screencoclorc <<EOF
screen mk kserver start
screen mk teststart
screen bash
EOF
    echo "Screenrc:"
    cat /tmp/screencoclorc
    screen -S coclo -c /tmp/screencoclorc
}
```