Use `container-vm` GCloud image
We've been having problems with the recent upgrade of the GKE
(cf. [release-notes](https://cloud.google.com/container-engine/release-notes)).

The problem seems to be the new VM image used by
[Google](https://cloud.google.com/container-engine/docs/node-image-migration)
which cannot mount NFS file-systems anymore.

This patch adds the option suggested at the
[end](https://cloud.google.com/container-engine/docs/node-image-migration#opting_out_of_using_container-vm_image)
of the documentation.
It has been tested and works like before (Kubernetes is 1.4 though), so we can
keep playing while we (or Google) figure out a longer term solution.