Add AWS-Batch support (issue #94)
The notion of cluster is an AWS Batch “job-queue,” Coclobas does not
create it but it checks that it's there; and fails at start-up if it
isn't “`VALID`”.

The `get_logs` function, does not et the logs, it instead generates a
URL to visit on <https://console.aws.amazon.com/cloudwatch>.

The tests now depend on (and use) `ppx_deriving_cmdliner`.

The plugin is for now very basic: it only runs commands and does not
understand `Program.t` values since there is no easy way of “mounting
data” on the containers.

When we go over-size for the command AWS-Batch fails with:

> An error occurred (ClientException) when calling the
> RegisterJobDefinition operation: Error executing request, Exception :
> Limits Error: JobDefinitionsize is above the limits 20 KB

For now, downstream users of the plugin should use other means to get
bigger scripts to run (e.g. use S3 or mount an EFS filer).

Killing jobs appears to work but it is somewhat fragile because AWS does
not really “confirm” the killing cf. issue
[`aws/aws-cli#2551`](https://github.com/aws/aws-cli/issues/2551)
(also, jobs take quite a while to change status in case of killing).