Use an optional S3 bucket to mount files
With the configuration option `--aws-s3-bucket` one can set an S3
prefix-URI to store files that the job will download.

For now, the S3 bucket must exist and have been made “semi-public” (with
`aws s3 website ...`). The scripts/files are passed as very hard to
guess URLs (“anyone with the link can see”); in the future we can give
the option of getting the files with `aws` but this requires having it
on the user-provided docker-image (with the right credentials).